Just use the ghc 7.8 cabal file.
Summary:
We are the haskell.org infra team, we can use modern tools.

I realize this might be contentious, but having two cabal files is annoying
because someone building on GHC 7.8 has to remember to `mv` the 7.8 file into
place but has to remember to move it back so they don't commit it on accident.

It is silly, let's just use 7.8 and get on with it.

Signed-off-by: Ricky Elrod <ricky@elrod.me>

Test Plan: `cabal install`

Reviewers: austin

Reviewed By: austin

Subscribers: relrod

Differential Revision: https://phabricator.haskell.org/D48